rename some files, share a dulpicated function
