+++ packages/web-workers/src/controllers/health-monitor-orchestration-controller.ts
@@ -8,8 +8,8 @@ import { inject, injectable } from 'inversify';
 import { ContextAwareLogger } from 'logger';
 import { WebController } from 'service-library';
 import { finalizerTestGroupName, getTestGroupClassNamesForScenario } from '../e2e-test-group-names';
+import { createScenarios } from '../e2e-test-scenarios/create-scenarios';
 import { E2EScanScenario } from '../e2e-test-scenarios/e2e-scan-scenario';
-import { createScenarios } from '../e2e-test-scenarios/e2e-scan-scenario-factory';
 import { OrchestrationSteps, OrchestrationStepsImpl } from '../orchestration-steps';
 import { WebApiConfig } from './web-api-config';
 // import { WebApiConfig } from './web-api-config';
